backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
google-generativeai==0.8.3
groq==0.11.0
mistralai==1.2.0
web3==7.4.0
py-solc-x==2.0.5
python-multipart==0.0.12
requests==2.32.0
pydantic==2.9.0
